Use non-build files
The forked project imports the project from an auto-generated build folder. Since we're not publishing to NPM and running a build script, this commit changes our imports to use the files themselves, rather than the build files.

The build files are identical, just copied to a `build` directory. See build.sh